+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ Unreleased
 Added
 -----
 
+- New algorithm for long-range corrected Hamiltonian
+
 
 Changed
 -------
@@ -18,6 +20,9 @@ Changed
 
 Fixed
 -----
+
+- Correct update of block mulliken population for onsite correction with
+  range-separation hybrid DFTB.
 
 
 19.1 (2019-07-01)
@@ -46,6 +51,7 @@ Added
 
 - Labelling of atomic orbital choices in output.
 
+- Halogen X correction.
 
 Changed
 -------